+++ src/main/java/com/example/whynotpc/persistence/img/ImageRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ImageRepo extends JpaRepository<Image, Integer> {
+public interface ImageRepo extends JpaRepository<Image, Long> {
     Optional<Image> findByName(String name);
 
     boolean existsByName(String name);
